Finish test suite setup
Have fixed the Guardfile so that the appropriate tests now run when a
change is made to a given file. Have also removed the yaml gem from the
Gemfile as this is part of Ruby's standard library and doesn't need to
be included.
Because the standalone_migrations gem uses ActiveRecord, migrations run
in the test environment require the `RAILS_ENV` variable to be set
first. Have added a note about this to INSTRUCTIONS.md.