Add #subtotal to Order model
Adding a #subtotal method to the Order model which will calculate the
subtotal before deals are applied. For now the method returns an integer
but currency formatting will be added in a later commit.

Have also had to alter the items factory slightly as it is easier to
add traits so that items with different attributes can be used in the
tests.