Replacing formatted subtotal with update methods
Formatting for the prices persisted in the database should probably take
place in the Checkout rather than the Order model itself so I have
removed this method. I will need to update the subtotal and total
persisted in the database in an after_save callback however so for now,
I have added `#update_subtotal!` and `#update_total!` to the model
instead. An order's total will be the same as the subtotal by default
though in later commits, the discounts applied should alter the total
accordingly.