Create the Checkout model
Have created the Checkout model which will not be persisted in the
database. At present, it is instantiated with GBP set as its currency by
default and can add new items to an order using the `#scan` method.
Later commits will add the ability to apply pricing rules.
Have also added tests for this class.