Rename the restrictions table
Have had a rethink about the architecture of the pricing rule
restrictions and I think the use of polymorphic tables would be better
than the simpler approach I initially had in mind. As such, this table
now makes more sense as the "restriction_types" table and the rows in
this table will correspond to the names of the new type-specific
restriction tables.