Create restriction_groups table
Added migration for the restriction_groups table. Currently the table
only has columns for the ID and timestamps though a later commit will
add a relationship to the PricingRule model. This model will be used to
group the different restrictions (e.g., minimum order value) that could
be set on a given pricing rule.