Create restrictions table
This migration will add the larger restriction table to the database. It
will have a belongs_to relationship with the restriction_groups and
restriction_types table. It also contains a column for specific
restrictions which will be associated (though not at the database level)
with records from any of the RestrictionType tables.