Create Restriction#body method
Similar to the method defined in the Reward model, I have added a method
that will return a row from the restriction type table depending on the
value of the restriction's `restriction_id` column.

Have also fixed issues with the `#type` and `#group` associations on the
model and refactored the related factories.