Add a RewardType::PricingRule#apply method
Have added a method that will apply rewards to a given order. This will
eventually be called by the PricingRule model and going forward, all
reward types will be expected to have such a method defined.
Have also added validation on the percentage to ensure that it isn't
possible to create "100% off" rewards.