Adding PricingRule#apply_reward
Have added the `#apply_reward` method to the PricingRule model so that
it can be used in the app to apply discounts. I also removed the
`#check_restrictions` method as I'm not sure that its use makes much
sense. The method itself was unlikely to be accessed by other classes
and turning it into a private method seemed to lead to duplication in
the `#apply_reward` method.
For now, rewards will be applied each time an order meets a restriction. I suspect
this could lead to unexpected results if a single pricing rule has wildly
different restrictions placed on it but this seems like the most
straightforward way to handle value off discounts on individual line items.