Added commands to the Supermarket CLI
Have added `init`, `display_product`, `scan`, `total` and `receipt`
commands to the CLI tool. It isn't possible to rely on access to
instance variables to keep track of changes to the database so I have
decided to store relevant data (such as the current order, existing
pricing rules and existing items) in a JSON file. Tests have been added
to ensure that each of the commands have the correct output though these
are likely to change in later commits.