Replacing the `apply` method in the Reward Types
Have decided to replace the `apply` methods with a `caluclate_deduction`
method in each of the reward types. Instead of directly applying a
deduction to a given order, the new method will caluclate the deduction
derived from a reward based on a given amount. The Value Off reward in
particular will use a parameter to determine how many times the
deduction can be applied though at present, it is unclear how such a
parameter could be used with the Percent Off reward type. The
application of this amount will be applied by the Checkout model in a
later commit.