Replace PricingRule#apply_reward method
Have replaced the `apply_reward` method with an `order_deduction` method
in the PricingRule as rewards should now be applied by the Checkout
model. This method will still check that all restrictions are met but
will now call the new `calculate_deduction` methods on the pricing
rule's reward. It will also use the number of matches to determine how
many times the deduction should be applied (which works well for single
restriction rules but could have unexpected results for rules with
multiple restrictions).