Add pricing_rules_applied column to orders
Have created a migration to add this column to the orders table so that
the app can keep track of deductions that have already been made. This
will be used in later commit to prevent the Checkout#total method from
applying more deductions than are necessary to a given order.