Apply rules when Order#pricing_rules_applied is false
Have refactored the Checkout#total method so that pricing rules are only
applied if the pricing_rules_applied column is set to false (which is
the default at the database level). It should then set this column to
true when this has been applied.  The Checkout#scan method will now set
this column to false again so that new deductions can be applied on
modified orders.