Handle pricing rules configured with missing rewards
Have refactored the PricingRule model so that it will handle rules that
have been set up without a reward. As the reward is an association, I
didn't think it made sense to add validation for this in the PricingRule
model so I will instead have to test for this edge case in the spec.