Add PricingRuleOrder model
Have added a model that corresponds with the pricing_rule_orders join
table. Have also added the necessary has_many relationships to the
PricingRule and Order models. The rules will be referred to as
"applied_discounts" on the Order model so that this can be included in a
final receipt for the user. The has_many on individual PricingRules
could, theoretically be used for reporting performed by a business analyst
(though this is out of scope).