Add applied discounts to orders in PricingRule
The PricingRule#discount_to_apply method should should create a row in
the pricing_rule_orders table whenever a discount has been applied to a
given order. If the order later becomes inelligible for a given
discount, this method should delete this row. Have run into a few issues
where orders with duplicate items are concerned, suggesting that there
could be an underlying issue with the has_many_through relationship.