Add Checkout#remove method
Have added a method that can be used to remove items from an order. Have
also cleaned up the logic of a pricing rule test that already did this
as the previous code would just return the deleted item rather than
removing it from a list (something that was missed because the list
consisted only of two instances of the same item). I'm having to convert
the ActiveRelation array into a standard Ruby array because ActiveRecord
doesn't seem to support deletion of a single instance of a duplicate
item.