Rename the INSTRUCTIONS.md file README.md
This project was originally written as part of a coding challenge for a
company who were unresponsive when I raised a PR. The original repo
contained a README with instructions for interview candidates which meant
that the instructions I wrote had to go in a separate file. As I've
dropped all of the commits that belonged to the company I was interviewing
for it should be safe to use the INSTRUCTIONS.md file as the README going
forwards.

On that note, I've since accepted an offer from a different company and as
such see no reason not to open source the work I have done here.